requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-groq==0.2.0
langchain-huggingface==0.1.0
pypdf
python-dotenv
streamlit
faiss-cpu